requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image Processing
pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0

# Deep Learning - PyTorch (for training)
torch>=2.1.0
torchvision>=0.16.0

# ONNX Runtime (for inference)
onnx>=1.15.0
onnxruntime>=1.16.0

# Utilities
pydantic>=2.5.0
